requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
groq>=0.4.0
aiohttp>=3.9.0
selenium>=4.15.0
playwright>=1.40.0
